refactor: add stories for pkgs
