+++ stories/components/LoadingIndicatorStories/Basic.tsx
@@ -1,0 +1,12 @@
+import {LoadingIndicator} from '../../../main';
+import {StoryContainer} from '../../GlobalStyles';
+
+function LoadingIndicatorSpinner(): React.ReactElement {
+  return (
+    <StoryContainer>
+      <LoadingIndicator />
+    </StoryContainer>
+  );
+}
+
+export default LoadingIndicatorSpinner;
