+++ stories/packages/PinchZoom/index.tsx
@@ -1,7 +1,6 @@
 import {ImageSlider} from './PinchZoomSlider';
-import React from 'react';
 import {storiesOf} from '@storybook/react-native';
 
-storiesOf('[Package] ImageSlider', module).add('Basic - light', () => (
+storiesOf('[Package] PinchZoom', module).add('ImageSlider', () => (
   <ImageSlider />
 ));
